Initial commit.
The missing sources and headers are automatically generated with STM32CubeMX and the `nosecone_fw.ioc` file.
The IMU data is fetched via DMA on in interrupt. the fetched payload is then decoded with the `Src/Sensors/IMU/imu.c:processBuffer()` function, that fills the `IMU_buffer` specified in `main.c`.